+++ src/App.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
+
 // Views _____________________________________
 import { Session } from "./views/Session";
 import { LandPage } from "./views/LandPage/LandPage";
@@ -13,6 +14,10 @@ import { Header } from "components/Header/Header";
 // Context ___________________________________
 import { UserContextProvider } from "./contexts/user";
 import { Footer } from "components/Footer/Footer";
+import { EventDetail } from "views/EventDetail";
+
+// Paths______________________________________
+const eventPath = [""];
 
 function App() {
   return (
@@ -25,6 +30,9 @@ function App() {
             <Route path="/login" element={<Session />} />
             <Route path="/main" element={<Main />} />
             <Route path="/events" element={<Events />} />
+            <Route path="/events/center/:id" element={<Events />} />
+            <Route path="/events/user/:id" element={<Events />} />
+            <Route path="/events/detail/:id" element={<EventDetail />} />
             <Route path="/*" element={<Error />} />
           </Routes>
         </main>
