+++ src/views/Events/index.jsx
@@ -1,6 +1,7 @@
+import { EventList } from "components/EventList";
 import Map from "components/Map/Map";
 import getEvents from "helpers/events/getEvents";
-import postEvent from "helpers/events/postEvent";
+// import postEvent from "helpers/events/postEvent";
 import { useSession } from "helpers/session/useSession";
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
@@ -35,21 +36,23 @@ const eventData = [
 export const Events = () => {
   const { jwt, user } = useSession();
   const [event, setEvent] = useState({});
-  const [showMap, setShowMap] = useState(false);
+  const [showMap, setShowMap] = useState(true);
   const navigate = useNavigate();
 
-  const handleCreate = async (e) => {
-    e.preventDefault();
-    setEvent((prevState) => {
-      let data = Object.assign({}, prevState);
-      data.author = user.id;
-      return data;
-    });
-    // validate all values
-    // This should be handled by map before selecting coordinates
-    await postEvent(event, jwt);
-    showMap(true);
-  };
+  // split("/events")[1] => NA, /user, /center
+
+  // const handleCreate = async (e) => {
+  //   e.preventDefault();
+  //   setEvent((prevState) => {
+  //     let data = Object.assign({}, prevState);
+  //     data.author = user.id;
+  //     return data;
+  //   });
+  //   // validate all values
+  //   // This should be handled by map before selecting coordinates
+  //   await postEvent(event, jwt);
+  //   showMap(true);
+  // };
 
   const handleSearch = async (e) => {
     e.preventDefault();
@@ -68,52 +71,49 @@ export const Events = () => {
 
   return (
     <>
-      {showMap ? (
-        <form onSubmit={(e) => e.preventDefault()}>
-          <input
-            type="text"
-            placeholder="sexo"
-            value={event.sex}
-            onChange={(e) => handleStateChange(e, "sex")}
-          />
-          {/* Dynamic optained from back ID */}
-          <input
-            type="text"
-            placeholder="deporte"
-            value={event.sex}
-            onChange={(e) => handleStateChange(e, "sportid")}
-          />
-          {/* ============================= */}
-          <input
-            type="text"
-            placeholder="accesibilidad"
-            value={event.sex}
-            onChange={(e) => handleStateChange(e, "accessibility")}
-          />
-          <input
-            type="text"
-            placeholder="nivel"
-            value={event.sex}
-            onChange={(e) => handleStateChange(e, "level")}
-          />
-          <input
-            type="text"
-            placeholder="disponibilidad"
-            value={event.sex}
-            onChange={(e) => handleStateChange(e, "date")}
-          />
-          <input
-            type="text"
-            placeholder="coste"
-            value={event.sex}
-            onChange={(e) => handleStateChange(e, "cost")}
-          />
-          <button onClick={handleSearch}>Search</button>
-          <button onClick={handleCreate}>Create</button>
-        </form>
-      ) : (
-        <Map data={eventData} />
-      )}
+      <form onSubmit={(e) => e.preventDefault()}>
+        <input
+          type="text"
+          placeholder="sexo"
+          value={event.sex}
+          onChange={(e) => handleStateChange(e, "sex")}
+        />
+        {/* Dynamic optained from back ID */}
+        <input
+          type="text"
+          placeholder="deporte"
+          value={event.sex}
+          onChange={(e) => handleStateChange(e, "sportid")}
+        />
+        {/* ============================= */}
+        <input
+          type="text"
+          placeholder="accesibilidad"
+          value={event.sex}
+          onChange={(e) => handleStateChange(e, "accessibility")}
+        />
+        <input
+          type="text"
+          placeholder="nivel"
+          value={event.sex}
+          onChange={(e) => handleStateChange(e, "level")}
+        />
+        <input
+          type="text"
+          placeholder="disponibilidad"
+          value={event.sex}
+          onChange={(e) => handleStateChange(e, "date")}
+        />
+        <input
+          type="text"
+          placeholder="coste"
+          value={event.sex}
+          onChange={(e) => handleStateChange(e, "cost")}
+        />
+        <button onClick={handleSearch}>Search</button>
+      </form>
+      <EventList eventData={eventData} />
+      <Map data={eventData} />
     </>
   );
 };
